requirements.txt
absl-py==2.1.0
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-RESTful==0.3.10
flatbuffers==24.3.7
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.28
tensorboard
tensorboard-data-server
tensorflow
termcolor
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
